refactor(organization-seat): simplify seat summary component and remove unused subscription cell
- Removed the organizationId prop from the OrganizationSeatSummaryCard component for cleaner usage.
- Replaced the MemberSubscriptionCell with a new SeatPlanCell component to streamline seat display logic in the members table.
- Deleted the MemberSubscriptionCell component as it is no longer needed.
- Updated related components and services to reflect these changes, enhancing code clarity and maintainability.
